Abstract cart/order components
