+++ src/components/header/MiniCart.vue
@@ -19,44 +19,22 @@
                 :key="lineItem.id"
                 data-test="mini-cart-line-item">
               <div>
-                <router-link :to="{
-                  name: 'product',
-                  params: { productSlug: lineItem.productSlug, sku: lineItem.variant.sku }}"
-                  class="img">
-                  <img :src="displayedImageUrl(lineItem.variant)"
-                       :alt="lineItem.name"
-                       class="img"/>
-                </router-link>
-                <router-link data-test="mini-cart-line-item-link"
-                             :to="{
-                             name: 'product',
-                             params: { productSlug: lineItem.productSlug, sku: lineItem.variant.sku }}">
-                  <p class="product-title">
-                    {{ lineItem.name }}
-                  </p>
-                </router-link>
+                <LineItemInfo :line-item="lineItem"
+                              :extended="false" />
                 <div class="details">
                   <p class="product-quantity">
                     {{ $t('quantity') }}
-                    <span data-test="mini-cart-line-item-quantity">
+                    <span data-test="cart-line-item-quantity">
                       {{ lineItem.quantity }}
                     </span>
                   </p>
                   <p class="product-price">
                     {{ $t('price') }}
-                    <span data-test="mini-cart-line-item-price">
+                    <span data-test="cart-line-item-price">
                       <BaseMoney :money="lineItem.totalPrice"/>
                     </span>
                   </p>
-                  <button @click="removeLineItem(lineItem.id)"
-                          data-test="mini-cart-line-item-delete"
-                          class="delete">
-                      <span>
-                        <img src="../../assets/img/delete-1.png"
-                             class="cart-action-icon"
-                             :alt="$t('delete')"/>
-                      </span>
-                  </button>
+                  <LineItemDeleteForm :lineItemId="lineItem.id"/>
                 </div>
               </div>
             </li>
@@ -72,7 +50,7 @@
                      class="btn-grey">
           {{ $t('viewBag') }}
         </router-link>
-        <router-link :to="{ name: 'cart' }"
+        <router-link :to="{ name: 'checkout' }"
                      class="btn-yellow">{{ $t('checkout') }}</router-link>
       </div>
     </transition>
@@ -88,9 +66,13 @@ import priceMixin from '@/mixins/priceMixin';
 import productMixin from '@/mixins/productMixin';
 import DisplayableMoneyFragment from '@/components/DisplayableMoney.gql';
 import BaseMoney from '../common/BaseMoney.vue';
+import LineItemInfo from '../common/cartlike/LineItemInfo.vue';
+import LineItemDeleteForm from '../cartdetail/LineItemDeleteForm.vue';
 
 export default {
   components: {
+    LineItemDeleteForm,
+    LineItemInfo,
     BaseMoney,
     VuePerfectScrollbar,
   },
@@ -105,27 +87,9 @@ export default {
     show() {
       return this.$store.state.miniCartOpen;
     },
-
-    sortedLineItems() {
-      if (this.me && this.me.activeCart) {
-        const { lineItems } = this.me.activeCart;
-        return lineItems.reverse();
-      }
-      return [];
-    },
   },
 
   methods: {
-    removeLineItem(lineItemId) {
-      return this.updateMyCart([
-        {
-          removeLineItem: {
-            lineItemId,
-          },
-        },
-      ]);
-    },
-
     toggle() {
       this.$store.dispatch('toggleMiniCart');
     },
@@ -154,7 +118,6 @@ export default {
           me {
             activeCart {
               id
-              version
               lineItems {
                 id
                 quantity
@@ -186,6 +149,12 @@ export default {
   },
 };
 </script>
+
+<style>
+  .nav-minicart .delete-text {
+    display: none;
+  }
+</style>
 
 <i18n>
   de:
