+++ tests/unit/specs/components/header/MiniCart.spec.js
@@ -6,22 +6,17 @@ const localVue = createLocalVue();
 localVue.use(Vuex);
 
 describe('MiniCart.vue', () => {
-  const lineItem = {
+  const lineItem = id => ({
+    id,
     quantity: 3,
     variant: {},
     totalPrice: {},
-  };
+  });
 
-  let me;
   let state;
   let options;
 
   beforeEach(() => {
-    me = {
-      activeCart: {
-        lineItems: [],
-      },
-    };
     state = { miniCartOpen: false };
 
     options = {
@@ -37,34 +32,6 @@ describe('MiniCart.vue', () => {
     expect(shallowMount(MiniCart, options).isVueInstance()).toBeTruthy();
   });
 
-  it('sorts line items', () => {
-    const wrapper = shallowMount(MiniCart, options);
-    wrapper.setData({ me });
-    expect(wrapper.vm.sortedLineItems).toEqual([]);
-
-    wrapper.setData({
-      me: {
-        activeCart: {
-          lineItems: [{ id: 'id1' }],
-        },
-      },
-    });
-    expect(wrapper.vm.sortedLineItems).toEqual([{ id: 'id1' }]);
-
-    wrapper.setData({
-      me: {
-        activeCart: {
-          lineItems: [{ id: 'id1' }, { id: 'id2' }, { id: 'id3' }],
-        },
-      },
-    });
-    expect(wrapper.vm.sortedLineItems).toEqual([
-      { id: 'id3' },
-      { id: 'id2' },
-      { id: 'id1' },
-    ]);
-  });
-
   it('close it according to store', () => {
     state.miniCartOpen = false;
     const wrapper = shallowMount(MiniCart, options);
@@ -75,7 +42,7 @@ describe('MiniCart.vue', () => {
     wrapper.setData({
       me: {
         activeCart: {
-          lineItems: [lineItem, lineItem, lineItem],
+          lineItems: [lineItem('id1'), lineItem('id2'), lineItem('id3')],
           totalPrice: {},
         },
       },
@@ -95,7 +62,7 @@ describe('MiniCart.vue', () => {
     wrapper.setData({
       me: {
         activeCart: {
-          lineItems: [lineItem, lineItem, lineItem],
+          lineItems: [lineItem('id1'), lineItem('id2'), lineItem('id3')],
           totalPrice: {},
         },
       },
